Move TextFilesUtils to text logic module
